Improve modularity of ign/ros publisher tests
